+++ sources/presentation/Stride.Core.Presentation/Collections/IObservableCollection.cs
@@ -1,0 +1,17 @@
+// Copyright (c) .NET Foundation and Contributors (https://dotnetfoundation.org/ & https://stride3d.net) and Silicon Studio Corp. (https://www.siliconstudio.co.jp)
+// Distributed under the MIT license. See the LICENSE.md file in the project root for more information.
+
+using System.Collections.Specialized;
+using System.ComponentModel;
+
+namespace Stride.Core.Presentation.Collections;
+
+/// <summary>
+/// This interface regroups the <see cref="ICollection{T}"/> interface, the
+/// <see cref="INotifyPropertyChanged"/> interface, and the <see cref="INotifyCollectionChanged"/>
+/// interface. It has no additional members.
+/// </summary>
+/// <typeparam name="T">The type of item contained in the collection.</typeparam>
+public interface IObservableCollection<T> : ICollection<T>, INotifyCollectionChanged, INotifyPropertyChanged
+{
+}
